deploy/ToraOps/requirements.txt
ansible==3.2.0
ansible-base==2.10.8
APScheduler==3.7.0
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.5
django-apscheduler==0.5.2
django-bootstrap==0.2.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-multiselectfield==0.1.12
djangorestframework==3.12.2
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
packaging==20.9
pandas==1.2.0
paramiko==2.7.2
Pillow==8.1.0
pip==21.0.1
pycparser==2.20
PyJWT==2.0.0
pymssql==2.1.5
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
setuptools==51.0.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
uWSGI==2.0.19.1
wheel==0.36.1
django-grappelli==2.15.1
coreapi==2.3.3